examples/requirements_common.txt
# Common
matplotlib==3.5.3
pandas
tensorflow_datasets
tensorflow_io
tensorflow_addons
imgaug
